+++ jabgui/src/main/java/org/jabref/gui/cleanup/CleanupFileViewModel.java
@@ -15,6 +15,7 @@ public class CleanupFileViewModel {
             CleanupPreferences.CleanupStep.RENAME_PDF,
             CleanupPreferences.CleanupStep.RENAME_PDF_ONLY_RELATIVE_PATHS,
             CleanupPreferences.CleanupStep.RENAME_PDF_ONLY_PDF_FILES,
+            CleanupPreferences.CleanupStep.RENAME_PDF_PRESERVE_SUFFIX,
             CleanupPreferences.CleanupStep.CLEAN_UP_UPGRADE_EXTERNAL_LINKS,
             CleanupPreferences.CleanupStep.CLEAN_UP_DELETED_LINKED_FILES,
             CleanupPreferences.CleanupStep.REMOVE_XMP_METADATA
@@ -25,6 +26,7 @@ public class CleanupFileViewModel {
     public final BooleanProperty renamePdfSelected = new SimpleBooleanProperty();
     public final BooleanProperty renamePdfOnlyRelativeSelected = new SimpleBooleanProperty();
     public final BooleanProperty renamePdfOnlyPdfFilesSelected = new SimpleBooleanProperty();
+    public final BooleanProperty renamePdfPreserveSuffixSelected = new SimpleBooleanProperty();
     public final BooleanProperty upgradeLinksSelected = new SimpleBooleanProperty();
     public final BooleanProperty deleteFilesSelected = new SimpleBooleanProperty();
     public final BooleanProperty removeXmpMetadataSelected = new SimpleBooleanProperty();
@@ -37,6 +39,7 @@ public class CleanupFileViewModel {
         renamePdfSelected.set(preferences.isActive(CleanupPreferences.CleanupStep.RENAME_PDF));
         renamePdfOnlyRelativeSelected.set(preferences.isActive(CleanupPreferences.CleanupStep.RENAME_PDF_ONLY_RELATIVE_PATHS));
         renamePdfOnlyPdfFilesSelected.set(preferences.isActive(CleanupPreferences.CleanupStep.RENAME_PDF_ONLY_PDF_FILES));
+        renamePdfPreserveSuffixSelected.set(preferences.isActive(CleanupPreferences.CleanupStep.RENAME_PDF_PRESERVE_SUFFIX));
         upgradeLinksSelected.set(preferences.isActive(CleanupPreferences.CleanupStep.CLEAN_UP_UPGRADE_EXTERNAL_LINKS));
         deleteFilesSelected.set(preferences.isActive(CleanupPreferences.CleanupStep.CLEAN_UP_DELETED_LINKED_FILES));
         removeXmpMetadataSelected.set(preferences.isActive(CleanupPreferences.CleanupStep.REMOVE_XMP_METADATA));
@@ -70,6 +73,9 @@ public class CleanupFileViewModel {
             if (!anyModifierSelected) {
                 activeJobs.add(CleanupPreferences.CleanupStep.RENAME_PDF);
             }
+        }
+        if (renamePdfPreserveSuffixSelected.get()) {
+            activeJobs.add(CleanupPreferences.CleanupStep.RENAME_PDF_PRESERVE_SUFFIX);
         }
         if (upgradeLinksSelected.get()) {
             activeJobs.add(CleanupPreferences.CleanupStep.CLEAN_UP_UPGRADE_EXTERNAL_LINKS);